Add machine-specific completion directories
I had to write fish shell code for the first time
and I actually really like it. It could be a good alternative
to xonsh if I get fed up with its quriks.

In our case, our macbook-specific completion files
are just links to the hut command installed in /opt/srht