Add SQLite-based cache to speed up config script
This allows us to avoid calling the slow `defaults` command,
with correct cache invalidation behavior if the file ever does change.

Dramatically speeds things up.
Before it was ~150-200 ms to call the `defaults` command.
We did it three times, which added large overhead.

Now takes ~70 ms to translate the macbook config.
Of that time, most of it seems to be python startup :/