common: Emulate ov-less on system with pacman
The AUR package for `ov` doesn't include an `ov-less` command.
However it doess include an `ov-less.yaml` config,
so we can easily emulate it.

My mac doesn't have this problem,
because it uses the Nix package for ov
installed via home-manager.

TODO: Avoid this problem by installing the Nix `ov` on both systems.