findrepo: For speed, prefer os.path to pathlib
This reduces average runtime from ~420ms to ~256ms, improving by 170ms.
Benchmarked by searching my ~/git directory on my M1 macbook.
Of course, runs are going to be much slower when the cache is cold.

The Path.join (and os.path.join) was taking up a lot of extra time.
It includes a lot of unecessary logic that we don't need.